+++ scalalib/src/mill/javalib/android/AndroidAppModule.scala
@@ -173,35 +173,12 @@ trait AndroidAppModule extends AndroidModule {
     Seq(androidSdkModule().androidJarPath())
   }
 
-  /** Generates a BuildConfig.java in the [[androidApplicationNamespace]] package
-   * TODO user configurable version code and version names and consider having a separate trait for this
-   */
-  def generatedBuildConfig: T[PathRef] = Task {
-    val buildType = if (androidIsDebug()) "debug" else "release"
-    val content: String =
-      s"""
-         |public final class BuildConfig {
-         |  public static final boolean DEBUG = ${androidIsDebug()};
-         |  public static final String APPLICATION_ID = "${androidApplicationId}";
-         |  public static final String BUILD_TYPE = "${buildType}";
-         |  public static final int VERSION_CODE = 1;
-         |  public static final String VERSION_NAME = "1.0";
-         |}
-      """.stripMargin
-
-    val destination = Task.dest / "source" / os.SubPath(androidApplicationNamespace.replace(".", "/")) / "BuildConfig.java"
-
-    os.write(destination, content, createFolders = true)
-
-    PathRef(destination)
-  }
-
   /**
    * Combines standard Java source directories with additional sources generated by
    * the Android resource generation step.
    */
   override def generatedSources: T[Seq[PathRef]] = Task {
-    androidLibsRClasses() ++ Seq(generatedBuildConfig())
+    androidLibsRClasses()
   }
 
   /**
